perbaiki profil desa

+++ app/Http/Livewire/Profil/EditPerangkatDesa.php
@@ -11,24 +11,37 @@ class EditPerangkatDesa extends Component
 {
     use WithFileUploads;
     public $perangkat_desa, $gambar;
+
+    public function mount()
+    {
+    }
     public function render()
     {
-        $this->gambar = ProfilDesa::find(1)->perangkat_desa;
+        $profil = ProfilDesa::first();
+        if (!$profil) {
+            $this->gambar = 'images/no-image.jpg';
+        } else {
+            $this->gambar = 'storage/' . $profil->perangkat_desa;
+        }
         return view('livewire.profil.edit-perangkat-desa');
     }
 
     public function update()
     {
-
-        $profil = ProfilDesa::find(1);
         $dataValid = $this->validate([
             'perangkat_desa' => 'required|image|max:2048',
         ]);
-        if (Storage::disk('public')->exists($profil->perangkat_desa)) {
-            Storage::disk('public')->delete($profil->perangkat_desa);
+        $profil = ProfilDesa::first();
+        if (!$profil) {
+            $dataValid['perangkat_desa'] = $this->perangkat_desa->store('images', 'public');
+            ProfilDesa::create($dataValid);
+        } else {
+            if (Storage::disk('public')->exists($profil->perangkat_desa)) {
+                Storage::disk('public')->delete($profil->perangkat_desa);
+            }
+            $dataValid['perangkat_desa'] = $this->perangkat_desa->store('images', 'public');
+            $profil->update($dataValid);
         }
-        $dataValid['perangkat_desa'] = $this->perangkat_desa->store('images', 'public');
-        $profil->update($dataValid);
 
         $this->dispatchBrowserEvent('swal:modal', [
             'type' => 'success',
